# docs/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisJob": {
        "additionalProperties": false,
        "properties": {
          "analysis_id": {
            "title": "Analysis Id",
            "type": "string"
          },
          "config_hash": {
            "title": "Config Hash",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "daily_episodes": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Anomaly episodes per trend day, oldest first; null for days without enough data. Present once a report is saved.",
            "title": "Daily Episodes"
          },
          "detector_version": {
            "title": "Detector Version",
            "type": "string"
          },
          "end_time": {
            "title": "End Time",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "explanation_status": {
            "$ref": "#/components/schemas/ExplanationStatus"
          },
          "finding_counts": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "propertyNames": {
                  "$ref": "#/components/schemas/Severity"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present once a report is saved.",
            "title": "Finding Counts"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "report_available": {
            "title": "Report Available",
            "type": "boolean"
          },
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageProgress"
            },
            "title": "Stages",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "trigger": {
            "$ref": "#/components/schemas/JobTrigger",
            "default": "manual",
            "description": "Submitted by a user or by the project schedule."
          }
        },
        "required": [
          "analysis_id",
          "scope",
          "end_time",
          "detector_version",
          "config_hash",
          "state",
          "stages",
          "explanation_status",
          "created_at",
          "report_available"
        ],
        "title": "AnalysisJob",
        "type": "object"
      },
      "AnalysisList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnalysisJob"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "AnalysisList",
        "type": "object"
      },
      "AnalysisReport": {
        "additionalProperties": false,
        "properties": {
          "analysis_id": {
            "title": "Analysis Id",
            "type": "string"
          },
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/MetricCapability"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "config_hash": {
            "title": "Config Hash",
            "type": "string"
          },
          "coverage": {
            "items": {
              "$ref": "#/components/schemas/SignalCoverage"
            },
            "title": "Coverage",
            "type": "array"
          },
          "detector_version": {
            "title": "Detector Version",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "exclusions": {
            "items": {
              "$ref": "#/components/schemas/Exclusion"
            },
            "title": "Exclusions",
            "type": "array"
          },
          "explanation": {
            "$ref": "#/components/schemas/ExplanationResult"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "schema_version": {
            "default": "2.1",
            "enum": [
              "2.0",
              "2.1"
            ],
            "title": "Schema Version",
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceInfo"
              },
              {
                "type": "null"
              }
            ],
            "description": "Deprecated: the first source; use sources."
          },
          "sources": {
            "description": "Every source analysed (2.1); empty in 2.0 reports.",
            "items": {
              "$ref": "#/components/schemas/SourceInfo"
            },
            "title": "Sources",
            "type": "array"
          },
          "state": {
            "$ref": "#/components/schemas/ReportState"
          },
          "trend_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrendSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "trends": {
            "items": {
              "$ref": "#/components/schemas/DailyTrend"
            },
            "maxItems": 14,
            "minItems": 14,
            "title": "Trends",
            "type": "array"
          },
          "windows": {
            "$ref": "#/components/schemas/AnalysisWindows"
          }
        },
        "required": [
          "analysis_id",
          "scope",
          "windows",
          "generated_at",
          "detector_version",
          "config_hash",
          "state",
          "capabilities",
          "coverage",
          "findings",
          "trends",
          "evidence",
          "exclusions",
          "explanation"
        ],
        "title": "AnalysisReport",
        "type": "object"
      },
      "AnalysisSubmission": {
        "additionalProperties": false,
        "description": "Body of POST /api/analyses. end_time is floored to the step server-side.",
        "properties": {
          "end_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id"
        ],
        "title": "AnalysisSubmission",
        "type": "object"
      },
      "AnalysisSubmitted": {
        "additionalProperties": false,
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/AnalysisJob"
          },
          "duplicate_of_active": {
            "default": false,
            "title": "Duplicate Of Active",
            "type": "boolean"
          }
        },
        "required": [
          "analysis"
        ],
        "title": "AnalysisSubmitted",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "validation_error",
          "project_not_found",
          "project_name_taken",
          "project_busy",
          "source_not_configured",
          "credentials_unreadable",
          "analysis_not_found",
          "report_not_ready",
          "report_unavailable",
          "schema_unsupported",
          "analysis_not_active",
          "queue_full",
          "metrics_source_unavailable",
          "end_time_invalid",
          "interrupted_by_restart",
          "job_timeout",
          "internal_error",
          "not_implemented"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
...
      "ExplanationStatus": {
        "enum": [
          "disabled",
          "not_configured",
          "pending",
          "succeeded",
          "failed",
          "skipped_no_findings"
        ],
        "title": "ExplanationStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "JobError",
        "type": "object"
      },
      "JobState": {
        "enum": [
          "queued",
          "running",
          "completed",
          "partial",
          "failed",
          "cancelled"
        ],
        "title": "JobState",
        "type": "string"
      },
      "JobTrigger": {
        "enum": [
          "manual",
          "scheduled"
        ],
        "title": "JobTrigger",
        "type": "string"
      },
...
      "MetricCapability": {
        "additionalProperties": false,
        "description": "Whether a signal family can be analysed for a scope, and why (not).",
        "properties": {
          "family": {
            "$ref": "#/components/schemas/SignalFamily"
          },
          "history_days": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Observed history for the scope; not proof of retention.",
            "title": "History Days"
          },
          "observed_metrics": {
            "items": {
              "type": "string"
            },
            "title": "Observed Metrics",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "required_labels": {
            "items": {
              "type": "string"
            },
            "title": "Required Labels",
            "type": "array"
          },
          "required_metrics": {
            "items": {
              "type": "string"
            },
            "title": "Required Metrics",
            "type": "array"
          },
          "signal": {
            "title": "Signal",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/SourceKind",
            "default": "prometheus"
          },
          "status": {
            "$ref": "#/components/schemas/CapabilityStatus"
          },
          "verified": {
            "description": "True only when observed in the live source.",
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "family",
          "signal",
          "status",
          "verified",
          "required_metrics",
          "required_labels"
        ],
        "title": "MetricCapability",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "description": "RFC 9457 problem details (application/problem+json).",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "ReportState": {
        "enum": [
          "completed",
          "partial"
        ],
        "title": "ReportState",
        "type": "string"
      },
...
      "Scope": {
        "additionalProperties": false,
        "description": "One project: every Prometheus query carries all of its matchers; reports stay inside it.\n\n``matchers`` is empty only for projects without a Prometheus source; Prometheus queries\nrefuse an empty scope.",
        "properties": {
          "matchers": {
            "items": {
              "$ref": "#/components/schemas/LabelMatcher"
            },
            "maxItems": 10,
            "title": "Matchers",
            "type": "array"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "project_name": {
            "title": "Project Name",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "project_name",
          "matchers"
        ],
        "title": "Scope",
        "type": "object"
      },
      "Severity": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Severity",
        "type": "string"
      },
...
      "StageName": {
        "enum": [
          "discovery",
          "collection",
          "detection",
          "trends",
          "explanation",
          "saving"
        ],
        "title": "StageName",
        "type": "string"
      },
      "StageProgress": {
        "additionalProperties": false,
        "properties": {
          "done": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Done"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "stage": {
            "$ref": "#/components/schemas/StageName"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/StageStatus"
          },
          "total": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "required": [
          "stage",
          "status"
        ],
        "title": "StageProgress",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DevOps AI Assistant",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/analyses": {
      "get": {
        "operationId": "list_analyses_api_analyses_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "List Analyses"
      },
      "post": {
        "operationId": "submit_analysis_api_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSubmitted"
                }
              }
            },
            "description": "OK"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSubmitted"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Submit Analysis"
      }
    },
    "/api/analyses/{analysis_id}": {
      "delete": {
        "operationId": "cancel_analysis_api_analyses__analysis_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Analysis"
      },
      "get": {
        "operationId": "get_analysis_api_analyses__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Analysis"
      }
    },
    "/api/analyses/{analysis_id}/report": {
      "get": {
        "operationId": "get_report_api_analyses__analysis_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report"
      }
    },
...
  }
}